pyproject.toml
[tool.poetry]
name = "zalgobot"
version = "0.1.0"
description = ""
authors = ["Avery Crespi <averycrespi@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
discord = "^1.0.1"
zalgo-text = "^0.6"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
